refactor(cli): flatten cli/__tests__ — rename 2 parallel suites (#2565 slice 6)
cli/__tests__ had 2 files with basename collisions:

- sandbox-exec.test.ts (top: 246 lines, general sandbox-exec tests)
  vs (nested: 544 lines, 8 describe blocks covering Input Validation /
  Command Allowlist / Command Injection Prevention / Path Traversal
  Prevention / Policy Selection / Safe Execution / Edge Cases /
  Security Regressions — dedicated security testing)

- session-storage.test.ts (top: 504 lines, vi.fn-based unit tests of
  SQLiteSessionStorage CRUD ops) vs (nested: 320 lines, class-based
  mock-database fake for SQLiteSessionStorage + Session storage
  helpers)

Rename nested files with scope-descriptive suffixes:
- sandbox-exec.test.ts → sandbox-exec-security.test.ts
- session-storage.test.ts → session-storage-mockdb.test.ts
  (consistent with the learning/outcome-storage-mockdb naming from
  PR #2579)

Fix `../X` static + `../../X` and `../X` dynamic imports to local
paths. Remove the empty __tests__/ directory.

All 3313 cli tests still pass.
